+++ changelog.md
@@ -1,3 +1,7 @@
+## 2.1.0
+- Added more feture for parsing value
+- Added the export function for exporting all the messages recived
+
 ## 2.0.0
 - Changed all the structure of the chart
 - Optimized the chart